change stack selection

+++ lib/bauble/cli/commands/destroy.rb
@@ -13,12 +13,17 @@ module Bauble
           def included(thor)
             thor.class_eval do
               desc 'destroy', 'Destroy the application'
+              method_option :stack, type: :string, desc: 'The stack to destroy', aliases: '-s'
 
               def destroy
-                @app.change_current_stack('dev')
+                raise 'No stacks found' if @app.stacks.empty?
+                raise 'Must provide a stack when multiple are defined' if @app.stacks.length > 1 && options[:stack].nil?
+
+                stack_name = options[:stack] || @app.stacks.first.name
+                @app.change_current_stack(stack_name)
                 pulumi.create_pulumi_yml(@app.template)
                 pulumi.init!
-                pulumi.create_or_select_stack('dev')
+                pulumi.create_or_select_stack(stack_name)
                 pulumi.destroy
                 Logger.log "Destroy complete\n"
               end
